FIX: lobby list 수정

See merge request s09-webmobile1-sub2/S09P12B109!213

+++ server/rhythm4cuts/src/main/java/com/b109/rhythm4cuts/model/repository/LobbyRepositoryImpl.java
@@ -32,6 +32,7 @@ public class LobbyRepositoryImpl implements LobbyRepository {
     @Override
     public GameInfo selectSeqLobby(int gameSeq) throws SQLException {
         // SELECT g FROM GameInfo AS g WHERE g.gameSeq = gameSeq
+        System.out.println("hihi");
         String jpql = "SELECT g FROM GameInfo g WHERE g.gameSeq = :gameSeq";
 
         return em.createQuery(jpql, GameInfo.class)
@@ -103,6 +104,8 @@ public class LobbyRepositoryImpl implements LobbyRepository {
     @Override
     public void deleteById(int gameSeq) throws SQLException {
         GameInfo gameInfo = em.find(GameInfo.class, gameSeq);
-        em.remove(gameInfo);
+        if (gameInfo != null) {
+            em.remove(gameInfo);
+        }
     }
 }
